backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
cohere>=5.1.4
pdfplumber==0.10.3
supabase>=2.10.0
python-multipart==0.0.6
pydantic==2.5.0
requests>=2.31.0
